feat(frontend): moved uniqueFieldNamesValidator and singleTrueValidator to validators module
